O-4_shape_matching_data-generator/requirements.txt
# Core dependencies
numpy==1.26.4
Pillow==10.4.0
pydantic==2.10.5

# Video generation
opencv-python==4.10.0.84
